+++ clients/admin/package.json
@@ -59,7 +59,9 @@
     "@tailwindcss/forms": "^0.2.1",
     "@tailwindcss/postcss7-compat": "^2.1.0",
     "@tailwindcss/typography": "^0.4.0",
+    "@testing-library/dom": "^7.31.0",
     "@testing-library/svelte": "^3.0.0",
+    "@testing-library/user-event": "^13.1.9",
     "@types/chai": "^4.2.13",
     "@types/mocha": "^8.2.0",
     "@types/node": "^15.3.1",
@@ -84,6 +86,7 @@
     "fromentries": "^1.3.2",
     "graphql": "^15.4.0",
     "graphql-tag": "^2.11.0",
+    "jest-in-case": "^1.0.2",
     "keycloak-js": "^12.0.4",
     "merge-deep": "^3.0.3",
     "mock-apollo-client": "^1.1.0",
