+++ clients/admin/src/absorb/smelte/src/components/Card/Title.svelte
@@ -1,5 +1,5 @@
 <script>
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
 
     export const hover = true;
     export let title = '';
