+++ clients/admin/src/absorb/smelte/src/components/DataTable/Editable.svelte
@@ -1,5 +1,5 @@
 <script>
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
     import { createEventDispatcher } from 'svelte';
     import TextField from '../TextField';
     import Icon from '../Icon';
