+++ clients/admin/src/absorb/smelte/src/components/DataTable/Row.svelte
@@ -1,5 +1,5 @@
 <script>
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
     import { createEventDispatcher } from 'svelte';
     import Editable from './Editable.svelte';
     import { Spacer } from '../Util';
