+++ clients/admin/src/absorb/smelte/src/components/NavigationDrawer/NavigationDrawer.svelte
@@ -3,7 +3,7 @@
     import { quadIn } from 'svelte/easing';
     import { Scrim } from '../Util';
     import breakpoints from '../../breakpoints';
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
 
     const bp = breakpoints();
