+++ clients/admin/src/absorb/smelte/src/components/RadioButton/RadioButton.svelte
@@ -2,7 +2,7 @@
     import Icon from '../Icon';
     import Ripple from '../Ripple';
     import Label from '../Checkbox/Label.svelte';
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
 
     const classesDefault = 'inline-flex block items-center mb-2 cursor-pointer z-0';
