+++ clients/admin/src/absorb/smelte/src/components/Ripple/Ripple.svelte
@@ -1,5 +1,5 @@
 <script>
-    import createRipple from './ripple';
+    import createRipple from '../../../../../dsl/ripple';
 
     export let color = 'primary';
     export let noHover = false;
