+++ clients/admin/src/absorb/smelte/src/components/Tabs/Tabs.svelte
@@ -1,6 +1,6 @@
 <script>
     import { onMount } from 'svelte';
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
 
     import Indicator from './Indicator.svelte';
     import ProgressLinear from '../ProgressLinear';
