+++ clients/admin/src/absorb/smelte/src/components/TextField/Label.svelte
@@ -1,5 +1,5 @@
 <script>
-    import utils, { ClassBuilder, filterProps } from '../../utils/classes.js';
+    import utils, { ClassBuilder, filterProps } from '../../../../../dsl/classes.js';
 
     export let focused = false;
     export let error = false;
