+++ clients/admin/src/absorb/smelte/src/components/TextField/Underline.svelte
@@ -1,5 +1,5 @@
 <script>
-    import utils, { ClassBuilder, filterProps } from '../../utils/classes.js';
+    import utils, { ClassBuilder, filterProps } from '../../../../../dsl/classes.js';
 
     export let noUnderline = false;
     export let outlined = false;
