+++ clients/admin/src/absorb/smelte/src/components/TextField/Wrapper.svelte
@@ -1,5 +1,5 @@
 <script>
-    import utils, { ClassBuilder, filterProps } from '../../utils/classes.js';
+    import utils, { ClassBuilder, filterProps } from '../../../../../dsl/classes.js';
 
     let classesDefault = 'mt-2 mb-6 relative text-gray-600';
