+++ clients/admin/src/absorb/smelte/src/components/Tooltip/Tooltip.svelte
@@ -1,6 +1,6 @@
 <script>
     import { scale, fade } from 'svelte/transition';
-    import { ClassBuilder } from '../../utils/classes.js';
+    import { ClassBuilder } from '../../../../../dsl/classes.js';
 
     const classesDefault =
         'tooltip whitespace-no-wrap text-xs absolute mt-2 bg-gray-600 text-gray-50 rounded md:px-2 md:py-2 py-4 px-3 z-30';
