+++ clients/admin/src/components/add-customer/AddOrEditCustomer.svelte
@@ -337,7 +337,7 @@
                                     id="legalAddressCountryId"
                                     label={$_('page.customers.add.country')}
                                     countryId={legalAddressCountryId}
-                                    form={$myForm}
+                                    form={myForm}
                                 />
                             </div>
