+++ clients/admin/src/dsl/Icon.svelte
@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
     export let small = false;
     export let xs = false;
     export let reverse = false;
