+++ clients/admin/src/dsl/classes.js
